* Repository chuyen het ve Integer thay vi Long ----> Check lai cac Service neu co van de
* Generate lai 3 file LoginInfo, UserInfo va Authorization
----> Doi ten cac ham Get Set ve dung chuan, co the gay ra conflict nhung chi can sua lai ten la xong

* Fix loi Register

* Add Logout
-----> Add dong nay vao de su dung  <a th:href ="@{/logout}"> Log Out </a>

+++ Mentor Finder/src/main/java/com/abc/WebApp2/repository/EnrolledRepository.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface EnrolledRepository extends JpaRepository<Enrolled, Long> {
+public interface EnrolledRepository extends JpaRepository<Enrolled, Integer> {
     
 }
